# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Appointments API",
    "description": "API for scheduling appointments between clients and service providers.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Use 'Bearer <token>' for JWT authentication"
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
        "/auth/register": {
          "post": {
            "tags": ["Authentication"],
            "summary": "Register a new user",
            "parameters": [
              {
                "name": "body",
                "in": "body",
                "schema": {
                  "$ref": "#/definitions/UserInput"
                }
              }
            ],
            "responses": {
              "201": { "description": "Created with token" },
              "400": { "description": "Invalid input" },
              "409": { "description": "Email already in use" }
            }
          }
        },
        "/auth/login": {
          "post": {
            "tags": ["Authentication"],
            "summary": "Authenticate a user and return a token",
            "parameters": [
              {
                "name": "body",
                "in": "body",
                "schema": {
                  "$ref": "#/definitions/LoginInput"
                }
              }
            ],
            "responses": {
              "200": { "description": "OK with token" },
              "400": { "description": "Missing email or password" },
              "401": { "description": "Invalid credentials" }
            }
          }
        },
        "/auth/google": {
          "post": {
            "tags": ["Authentication"],
            "summary": "Authenticate via Google idToken",
            "parameters": [
              {
                "name": "body",
                "in": "body",
                "schema": {
                  "$ref": "#/definitions/GoogleInput"
                }
              }
            ],
            "responses": {
              "200": { "description": "OK with token" },
              "400": { "description": "Missing or invalid idToken" }
            }
          }
        },
...
      },
      "definitions": {
...
    "UserInput": {
      "type": "object",
      "properties": {
        "authProvider": { "type": "string", "example": "google" },
        "authId": { "type": "string", "example": "1092837498127398" },
        "email": { "type": "string", "example": "user@example.com" },
        "password": { "type": "string", "example": "plaintext" },
        "name": { "type": "string", "example": "John Doe" },
        "phone": { "type": "string", "example": "+1234567890" },
        "role": { "type": "string", "example": "client" },
        "avatarUrl": { "type": "string", "example": "https://example.com/avatar.png" }
      }
    },
...
        "LoginInput": {
          "type": "object",
          "properties": {
            "email": { "type": "string", "example": "user@example.com" },
            "password": { "type": "string", "example": "password123" }
          }
        },
        "GoogleInput": {
          "type": "object",
          "properties": {
            "idToken": { "type": "string", "example": "GOOGLE_ID_TOKEN" }
          }
        }
      }
}